LPS-77573 Move constants to api module
