+++ modules/apps/web-experience/product-navigation/product-navigation-product-menu-api/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay Product Navigation Product Menu API
+Bundle-SymbolicName: com.liferay.product.navigation.product.menu.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.product.navigation.product.menu.constants
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title: Product Navigation
