+++ modules/apps/web-experience/product-navigation/product-navigation-product-menu-api/src/main/java/com/liferay/product/navigation/product/menu/constants/ProductNavigationProductMenuWebKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.product.navigation.product.menu.web.constants;
+package com.liferay.product.navigation.product.menu.constants;
 
 /**
  * @author Julio Camarero
