+++ modules/apps/web-experience/product-navigation/product-navigation-product-menu-api/src/main/resources/com/liferay/product/navigation/product/menu/constants/packageinfo
