+++ modules/apps/web-experience/product-navigation/product-navigation-product-menu-theme-contributor/src/main/java/com/liferay/product/navigation/product/menu/theme/contributor/internal/ProductMenuTemplateContextContributor.java
@@ -21,7 +21,7 @@ import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.SessionClicks;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.product.navigation.product.menu.web.constants.ProductNavigationProductMenuWebKeys;
+import com.liferay.product.navigation.product.menu.constants.ProductNavigationProductMenuWebKeys;
 
 import java.util.Map;
