+++ modules/apps/web-experience/product-navigation/product-navigation-product-menu-web/bnd.bnd
@@ -1,9 +1,6 @@
 Bundle-Name: Liferay Product Navigation Product Menu Web
 Bundle-SymbolicName: com.liferay.product.navigation.product.menu.web
-Bundle-Version: 2.0.23
-Export-Package:\
-	com.liferay.product.navigation.product.menu.web.constants,\
-	com.liferay.product.navigation.product.menu.web.display.context
+Bundle-Version: 3.0.0
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Product Navigation
 Web-ContextPath: /product-navigation-product-menu-web
